chore(database): move entities to domain folder and include specialist
