+++ src/app/database/database.module.ts
@@ -1,14 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Patient } from '@/domain/entities/patient';
-import { PatientSupport } from '@/domain/entities/patient-support';
-import { Token } from '@/domain/entities/token';
-import { User } from '@/domain/entities/user';
+import { DATABASE_ENTITIES } from '@/domain/entities/database';
 import { EnvModule } from '@/env/env.module';
 import { EnvService } from '@/env/env.service';
-
-export const DATABASE_ENTITIES = [User, Token, Patient, PatientSupport];
 
 @Module({
   imports: [
