+++ src/domain/entities/database.ts
@@ -1,0 +1,13 @@
+import { Patient } from './patient';
+import { PatientSupport } from './patient-support';
+import { Specialist } from './specialist';
+import { Token } from './token';
+import { User } from './user';
+
+export const DATABASE_ENTITIES = [
+  User,
+  Token,
+  Patient,
+  PatientSupport,
+  Specialist,
+];
